Initial commit
This initial version already has a large number of features, a incomplete list being;

A collection of scripts to define data and behaviors of objects in the game including Unit, Physical, Buildable, Mortal, Motile, Builder and Attacker. These interact in a lightweight way to allow composition of many different types of objects quickly.

A user interface mostly built on top of a "Drawers", containers of "Tiles" (essentially square UI primitives) that display their contents in a grid using the available space. They can also be given handles, and thus be dragged to show more or less of their contents (thus the name "Drawer").

A custom physics system that treats physics elements as particles. This is intended to either evolve into a system capable of handling a very large number of physical elements, or ease the integration and inform the design of such a system created as a separate project.

"Variable"s; interfaces that allow various classes to emit named values for display in the UI and interacted with by the player. For example, the Mortal script has a "Health" Variable that is displayed in the Unit VariableDrawer.

Stub classes and UI elements for the future implementation of a advanced command queue which will enable players not just to issue a series of commands, but to visually edit the queue in order to repeat sets of commands, execute different command based on Variable values, and write to user variables to enable algorithmic state.

Class design that enables nearly everything to be controlled through the unity editor. For example, there is a "Unit" prefab from which a variant can be made that fully defines a completely new type of unit, complete with complex behaviors.

Establishing the visual style of the game; very dark background colors and bright, saturated, thin and pixelated details.