Adding Memory class
Memory is for storing values not inherent to the function of a class (like health is for Mortal), making them retrievable as Variables. This is especially useful for readonly values (like storing a position when the user clicks the map) and writable values of arbitrary type (like when the user needs to store an int/string/etc). The emphasis here is values that are created and/or modified at runtime, based on user input.