Operation input, output
This commit adds the capability for Operations to take Variables as input and/or output, as well as take Operations as a sort of input to special Operations (ChooseOperation only right now). This standardization of the interface makes visualization simpler as well simplifying the game mechanics for the user.

Variable inputs and outputs are handled by a new class "VariablePipe" which takes a name and then finds the variable with that name (contextual to the unit executing an Operation) and allows reading and/or writing to that variable. This allows for complex arbitrary interactions between Variables and Operations in away that doesn't require a lot of state validation (Worst case scenario is the variable name is changed or collided, which results in predictable and preventable effects).

VariablePipe also has initial support for the future feature of applying unary and binary functions to inputs and outputs. This means that VariablePipe also contains an optional secondary variable name and a "PipeFunction" which can take 1 or 2 values as input to produce a single value as the final value in a Read() or Write().

Tasks, which under the current design are universally parameterized by a position/gameobject, now have a Target property which reads their input and attempts to extract a Target from it. To support this, Task now also has a Unit property (to give context to the variable name stored in Operation.Input).