Limbo
Added new class "Limbo". This class gathers up Mortals with 0 or less health and removes them from the World. Currently, it just stores them after that.

This class is intended to help with scenarios where an object needs time to settle its affairs, or support any feature relating to death.