_3DUIElement
Added classes _3DUIElement and _3DUIElementContainer. These classes support in-world elements that belong to 2D ui elements. Any object with _3DUIElement attached will, on Start(), change its parent to the _3DUIElementContainer in the scene.

This avoids transform problems that can arise when a ordinary Transform has a parent RectTransform. In particular, it appears that the z coordinate is locked under these circumstances.