Left mouse button release event claiming
Added ability for MonoBehaviours to "claim" the next left mouse button release event. This is done through InputUtility.

This functionality helps whenever a left mouse click is expected as part of a use case and you don't want other classes to misinterpret that click. For example, the user has indicated they are selecting a OperationTile, but OperationTile normally reacts to being clicked. By Claiming the event, OperationTile can ignore that click.

This functionality would work in dragging scenarios, but because MainInputModule.DidDragOccur universally captures any drag event without additional code, it is preferred in those situations.

Extending this functionality to other events, in particular the right mouse release event, is not currently needed so I'm leaving it out for now.