OperationTile dragging changes
Made OperationTile's only draggable when they are in the ProgramInterface, added code to trigger insertion previews when dragging, and added ability for user to remove an Operation by dragging it's Tile off the ProgramInterface and letting go. This last use case also comes with a red x overlay to indicate the user is about to remove the Operation.