OperationTile bugfix
Fixed bug where an OperationTile being dragged tried to find Unit's team color. In this scenario, it doesn't have a unit, because that is derived from the UnitInterface, which it isn't a part of anymore.

Just made Unit property return null when this happens and disabled color changes if Unit == null.