SurfaceDeposit
Added new sublclass of Deposit "SurfaceDeposit".

SurfaceDeposits have a .MaxVolume that is always greater or equal to .Volume. As material is added, the max volume increases, and its .Extent increases as well (proportional to the cube root of MaxVolume).

SurfaceDeposits can be created and added to by using new method CelestialBody.Litter() which adds resources to nearest overlapping SurfaceDeposit, or creates one if none such exist.

CelestialBody also checks whether two deposits are overlapping and merges them, with the larger deposit absorbing the smaller.

Finally, the prefab for SurfaceDeposit visualizes it as a conical pile
on the surface of the terrain. When .Volume < .MaxVolume, it scales the
pile in the y axis in proportion to the fill ratio.