SerializableDictionary DictionaryPropertyDrawer bugfix
This class was causing a null reference error when adding elements to a list containing SerializableDictionaries in the editor.

Essentially, what was happening was an element was added to the list (through the editor). Somehow, the editor "remembers" whether a SerializableDictionary should be visually expanded or not based on its variable name (based on the last state that name had). If the state was expanded, it asks the DictionaryPropertyDrawer to give a height for its visual element, but at this point, the list is still empty (the editor hasn't actually added the new element yet). This seems to violate this class's assumptions, resulting in invalid state and subsequent attempt to access a null reference.

I've simply added an if statement that detects this invalid state and returns 0 height.