Refiner
Added class for units that take regolith/surface desposits and separate out particular resources, producing a pile with those resources enriched, and another, depleted.

Its capabilities are determined a list of RefineTasks it contains. RefineTasks describe the "affinities" this refinement process has, i.e. what resources are attracted to the output pile and to what degree they are attracted. It also defines the efficiency of the separation, meaning how much unrelated material winds up in the output pile.

RefineTask repurposes the Output VariablePipe for what is, in a technical sense, an input; the location of the refined pile. This however is very intuitive for the user, as from their perspective, the input is the material coming in and the output is the material coming out of the refinery.

Abuse of the internal logic by the visual layer is, I believe, very natural, and overall, for the best in cases like this. As long as something is intuitive to the user, it should be supported and code should be added to interpret their actions in context.