ConvertBehavior explicit ConvertTask parameterization
ConvertBehavior now suppports being spawned without Unit.Task being a ConvertTask.

Because it would result in code duplication otherwise, it still uses a ConvertTask to parameterize it, but now that ConvertTask is a public property rather than being access through Unit.Task.

Most behaviors that access Unit.Task should probably be updated to work like this, as their current implementation prevents behavior composition, but I don't yet have a need for that, so I'm leaving them as is for now.