Factory prefab support
The Factory prefab is intended to build units on top of itself. this however is not supported because

1) When being built, a Motile is considered as part of the same collision layer as buildings, resulting in the unit immediately being forced off the factory.

2) BuildTask's normally take an input location, but this is not desired as the factory has a fixed construction site.

3) Once built, Units cannot be selected because the factory's hitbox is larger than theirs.

I've address these issues by 1) making Motiles under construction remain in the Motile collision layer, 2) adding a fixed construction site option to both Builder and BuildTask, and 3) reworking the Unit selection test to account for the situation where a Motile is ontop of a non-Motile. That last one may seem a bit specific, but I only allow intersections between Motiles and non-Motiles so it actually seems appropriate.