MathUtility easing functions
Added "ZenoLerp" and "ZenoLerpAngle".

I very commonly use Mathf.Lerp to achieve simple easing that doesn't require keeping track of where and when the easing began by simply lerping a fraction of the way from current to target value (proportional to frametime).

The problem is that this never actually reaches the target value. To fix this, ZenoLerp also takes a distance that defines at what scale this fractional stepping comes to an end. Currently, this is implemented as a gradual increase in relative lerp speed to roughly 10x as it approaches the critical distance, and further speedups as it gets even closer than that.