RailgunCrawler
Added a ranged projectile Unit "RailgunCrawler".

This comes with the addition of the class "RangedAttacker", its subclass "ProjectileAttacker", and some changes to Attacker.

Attacker now has a boolean state "IsAttacking".

RangedAttacker is a subclass of Attacker. It assumes the existance of some kind of barrel/muzzle that can be aimed up and down only. It provides two behaviors: RangedAttackBehavior and AimBehavior.

RangedAttackBehavior spawns an AimBehavior and waits for the greenlight from that class to tell the RangedAttacker to start firing. AimBehavior is able to calculate the direction to turn and the angle to fire at in order to hit a target given a projectile velocity, and indicate when aiming is complete.

To implement the attack, RangedAttacker simply calls Fire() (an abstract method) every RangedAttacker.Cooldown seconds. It also takes care of resource costs by calling abstract method GetCartridgeRequirements() which subclasses implement to describe how many and what resources are required to fire a single shot. RangedAttacker creates local storage space using Hoarder equal to RangedAttacker.MagazineSize * GetCartridgeRequirements(), and then creates a ongoing request to fill that storage space.

ProjectileAttacker implements RangedAttacker's by Fire()ing a Projectile along the FiringDirection with specified ProjectileVelocity.

Projectile is a new class that represents an object moving through space while under the effects of gravity. It imparts energy on objects it intersects, both as kinetic energy, and destructive (damage dealing) energy. The ratio of damage to velocity change is determine by Mortal's new structural fields ImpactFragility and ImpactTolerance, as well as Projectile's Penetration field.

Finally, these new classes are all used to make the new prefab "RailgunCrawler". This is a mobile unit that fires heavy projectiles at high speed.