Unit targets via mouse selection
User can now select an on-screen unit as the target of f.e. an attack command.

Unit's now have unique identifiers, which map to references to that unit through World.Memory. When a unit is clicked on during a OperationTileIONode selection use case, their identifier is used as the variable name connected to that node.