Buildable material changes
Instead of having required materials and retained materials, now Buildable has "structural" materials and "process" materials. structural materials are equivalent to the earlier retained materials; they are those resources that the Buildable is actually made out of. Process materials are the resources that are consumed during construction, like energy and tools.

Required materials still exist, but are now a property that simply adds structural materials to process materials.

The information is still the same (process materials are just required materials minus retained materials). But this is more constructive and is not capable of having invalid state (you could retain materials not in the required materials).