Physics optimization
Profiling determined that Physics was spending a lot of time calling Monobehaviour.HasComponent<>() in order to determine if a Physical was Motile or Projectile.

I've replaced this a concept of Physics "layers". Moving units are now part of the "Vehicle" layer, bullets are in the "Projectile" layer, etc.

These layers are defined only once in the Start() method of Physical, so the instigating performance issue is now fixed.