Variable.Read<T>()
Added typed Read() method to Variable.

This is just syntactic sugar if you know the type of the value, but in the case where you don't know the type (useful for WritableVariables), or don't know whether the variable exists, it also provides the ability to return a default value of your choosing.