ApplianceBehavior.IsProducingTools
Because running out of tools was unrecoverable (Foundries require tools to make tools), I added a boolean to ApplianceBehavior for subclasses that are producing tools. If true, ApplianceBehavior stops requesting tools.

To complete this new functionality, subclasses that make use of this feature have to subtract their ordinary tool request from their output. Currently, I'm just allowing the two relevant subclasses to duplicate eachother's implementation of this functionality, as no worthwhile way occurs to me to abstract it.

It gets sticky when the tool producer is making less tools than it is producing, as it can cause a deficit (negative number of tools) if tools run to zero. To avoid this, negative tool producers have their Unit.Task set to null.

This is a rather messy implementation, one because there is no formal contract that tool producers have to fulfill, and two because behaviors don't necessarily share a lifecycle with the current Task (meaning setting it to null might do nothing).

This reflects how messy the model is - its realistic that running out of tools you use to make tools would result in a real problem, though this would rarely actually happen, and wouldn't happen to all tool producers simultaniously.

I considered many ways to emulate how this problem would be avoided in real life, (such as keeping a local stock of tools), but all the solutions were either incomplete (i.e. edge cases could cause the problem to happen anyways), overly complex, opaque to the user, or all three.

Eventually, I'd like to iterate on this solution.