Refiner cycling
Before, it was sometimes advantageous to process a Refiner's output again, to raise its purity further. This was very clunky to set up in game. I decided to streamline this functionality by providing a "Cycles" WritableVariables that defines how many times to reprocess a sample before outputting it.

This is actually much more complicated than it sounds, if you want to do it analytically rather than simply iterating the refining code n times.

Because of this complexity, I decided to get rid of the "SeparationEfficiency" and replace it with "GangueAffinity" (representing the affinity of everything not in RefineTask.Affinities). By treating each component as an independent affinity, I can compute the final volume of each by a simply multiplying the initial volume by its affinity raised to Cycles.

One consequence of easy reprocessing is that 99.999% purity of the principle resource can be achieved with enough cycles, regardless of the parameters. You could adjust the numbers to make high purity overly expensive, but not without losing quite a bit of power to describe different refinements. Besides that, its unrealistic, as every refining/beneficiation process is going to have an upper limit of purity.

So, I added a MaximumConcentration field to RefineTask. I wanted to make this maximum require infinite cycles to reach (i.e. I wanted slowly dimishing returns as you approach max), but this would require integration, and I decided it wasn't worth the effort. Instead, I just compute the cycles required to reach maximum concentration, and then make that the maximum number of cycles allowed.