MainScene organization
MainScene was becoming a real hassle to untangle during the commit process. I decided the simplest and best solution was cut the scene up into a number of prefabs.

Some are mostly expedient, such as the Style prefab, which I just want so I have less text to look through during untangling. These I've put in the new Prefab folder "Globals" meaning their are singletons that will be used unchanged in every scene. I suppose they are necessary in the longterm anyways, as if I create new scenes, I'll need to link those assets into them.

Others may prove to be very useful for instancing and mixing and matching, such as the new folder for CelestialBodies, which I in theory use to quickly create a World of several Asteroids, for example, or Scenarios, which I put Worlds into that could wholly encapsulate a mission in the game.