UnitLoadUnloadVisualizationController.Start() bugfix
Before the lines and circles this class uses are separated from the Unit's they come from, they get scaled if the Unit has a scale. They are supposed to be unscaled, so I've added code to ensure this.

I've gone ahead and refactored the code here as well because with this addition, the code duplication exceeded what I thought was appropriate.