InfoBox
Added new class and prefab InfoBox, which displays mouse position contextual information in a small window below the mouse pointer.

InfoBox determines if the user is pointing at an object that has a component that implements the HasInfos interface, and if so, uses that to display information to the user about that object.