Variable.Style
Variables no longer have a Color and Sprite themselves, and thus no longer need to be .Stylized().

Instead, they have a .Style property which looks up their style in Scene.Main.Style.

This system was already in place indirectly, requiring every Variable returner to individually stylize their variable. This was intended to be open-ended and to allow data driven styling, but as the requirements have solidified, I no longer think this is necessary.