OperationTileIONode.VariableTile changes
OperationTileNode no longer actually stores a reference to the VariableTile it points to, instead just storing the name of the Variable, and looking up the VariableTile upon request.

This makes it more flexible, as the named Variable doesn't have to be available in a VariableTile immediately.

In particular, this was causing a race condition in the position selection use case, necessitating a forced update to World.Variables. I've removed that fix now.