UnitInterface prefab : ProgramInterface handle, next OperationTile
Added Handle to ProgramInterface

Added an OperationTile that represents the Unit.Program.Next Operation.

When ProgramInterface closed, if there is a Unit.Program.Next Operation, the new OperationTile slides out from the "UnitInfo" element. When open, it slides back in.

To make space when sliding out, ProgramInterface is moved down.