VariablePipe : removing identity PipeFunction
A blank PipeFunction is now null instead of the identity function. Mainly, the issue is that visually I would have to hide the identity function because it doesn't add anything to the user experience except possibly confusion.

I generally try to keep the logic and the visuals in sync with each other as much as possible.