Pathing
Added several new classes to for pathing.

Graph - Graphs are a collections of Nodes. Nodes contain two pieces of data : a list of neighboring nodes, and the cost to move to each of these neighbors. With this information, the "edges" in the graph can be defined.  Nodes also have the ability to estimate the path cost to nodes that aren't their neighbor.

Nodes have a method .GetPathTo(Node other), which implements the A* pathfinding algorithm.

Nodes are "implemented" by passing them a object of type IData which requires implementers to provide the information necessary for full Graph definition as well as cost heuristics for pathing.

Graph has many methods for creating and editing Graphs, including

1) CreateGrid() which makes a xz plane 2d square grid Graph with specified resolution and size. Diagonal edges are included. A overload of this method exists to pull a grid over a Terrain object (taking on y heights of terrain).

2) CreateHairball() which simply takes a set of points and connects every node to every other node.

3) Difference() and Union() for Graph set operations

4) Defleshed() which is like an set intersection, but first involves mapping one graph onto another. Essentially, paths are found in Graph A to travel along edges in Graph B, and then any nodes in A that are not along those paths are removed.

5) MinimumSpanned() implements Primm's alogorithm for finding the minimum spanning tree of a Graph.

Finally, added a class "HighwaySystem" that makes use of all these tools to produce a heightmapped grid Graph of the Asteroid, a minimum spanning tree of the major locations on it (right now uses locations of resource Deposits), then essentially creates a number of two lane highways through the grid along the minimum spanning tree. By calling PlanRoadtrip(), callers can recieve a detailed, nearly optimal path through the map that avoids collisions with vehicles traveling in the opposite direction.

To implement this new system, changes were made SeekBehavior to make it request a path from Asteroid.HighwaySystem and then follow it quickly by estimate their progress through the path and trying to drive towards the node x hops ahead of their current position (currently that x is 4). Eventually, I'd like to compute this x value, as well as tighten how closely vehicles stick to the path while still traveling at high speed.

Currently, edge costs are purely the straigh line distance. Future commits will add more sophisticated cost metrics, such as the incline of the ground, avoiding buildings, etc.