NodeExtensions.GetPathTo() "guest" destination
Added ability to use a "guest" destination - a node that isn't actually connected the source node. In this case, the system considers every node to have the destination as a neighbor.

If the graph represents all possible actions that can take place, this feature is useless. However, if the graph is merely a simplification of a more complex decision, this allows you to leverage the pathfinding algorithm to integrate its model with the real world by comparing, at every point in the path, the costs of following the model to the alternative of just going straight to the destination.

In my particular use case, the highway graph is strongly directional and designed to maxamize bidirectional throughput of units to a relatively small number of important sites on the map. This means that units may go out of their way to get on the highway unecessarily, and, harder to solve, they may follow the highway further than necessary (because the nearest highway node may be going the opposite direction, necessating a u turn at one of the infrequent turnoffs). With guest destinations, the unit will straight line to the destination the moment it is optimal.

This doesn't resolve interior graph suboptimalities (imagine a road the doubles back on itself but is inside a mountain range so the destination is still a long way off - the unit will still have to follow the road despite the possibility of straight line between crest of the road possibly being better), however this really can only be address by making the graph better because it defeats the purpose of having a simple model to then turn it into a hairball.

It also doesn't adress how to optimally enter the graph, what you would call a "guest source", which can be solved similarly by connecting the source node to all nodes in the graph. A future commit adds this functionality.