GraphUtility.CreateHeightMetric(), ObstacleMetric changes
CreateHeightMetric() now uses radian angles to measure slope instead of rise/run, and allows custom functions to compute the angle of a slope.

The latter is necessary because the slope of the path is almost always an underestimate of the actual slope of the surface because you may be climbing the slope obliquely. To calculate it accurately some sort of gradient computation is necessary. For the Terrain however, unity already provides a method to compute the gradient, so I'm now using that to compute the slope.

ObstacleMetric, in addition to making use of the custom slope option, is now also using the different formula to determine the cost of the slope.