SeekBehavior changes
Seekbehavior had a problem with paths with large edges because it just found the nearest node in the path and went to node after that (With progress state to avoid going backwards).

Now, it attempts to look ahead spatially a certain distance, and if that lead is beyond the distance to the next node, it updates its progress to the first node that is beyond the lead. Finally, it goes the node after its current progress (progress here representing which nodes you've visited already).