HighwaySystem actual left and right lanes
Before there were "left" and "right" edges/lanes/highways etc, but it was an arbitrary distinction. Now, forward travel is always to the right of opposing traffic.

This is accomplised by arbitrarily selecting a "capital" city node that only has one neighbor (its a dead end on the MST) and then finding all paths extending out from it. Using these edges as the left lanes and the opposite edges as the right lanes, the right lane's nodes are offset along the edge direction crossed by the up vector.

This helps make the right lane offsets more consistant, resulting in less crossings/intersections. It may also help in the future with collision avoidance because vehicles can move further right to increase the margin with opposing traffic.